third_party/WebKit/LayoutTests/fast/spatial-navigation/snav-container-select-expected.txt

startexpected targetunexpected target
PASS successfullyParsed is true

TEST COMPLETE
PASS gFocusedDocument.activeElement.getAttribute("id") is "expected"
